# btf: don't copy types during CO-RE relocation